api/pyproject.toml
[project]
name = "serfor-api"
version = "1.0.0"
description = "SERFOR Natural Language Query API with AI Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "instantneo[openai]>=0.1.1",
    "python-dotenv>=1.0.0",
    "pyodbc>=4.0.39",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
